Changed keys of _values array to attribute names.
Changed the values of the _values array to be the attribute names of the
RiverModule class. Since the output arrays of RiverModule change with
every time step, we can't just get pointers to the initial arrays and
keep those. Instead we need to get new arrays every time we call
get_values.

Convert lists returned by the RiverModule getters to numpy arrays.

Removed unused setter for shoreline.

Removed getter for river_mouth_location as it didn't seem to have a
corresponding attribute in the RiverModule class.